💄 style: Update ignore files for canisminor1990
This commit updates the .eslintignore, .gitignore, .npmrc, and .prettierignore files to better ignore files and folders that are not relevant to the codebase. The changes include ignoring dependencies such as node_modules, adding patterns for ci related files such as .coverage, and refining the patterns for production files such as dist and logs. Additionally, this commit adds new patterns such as ignore files (.DS_Store) and Dockerfiles.